feat(core): extract a plain cloud thread list adapter factory
