+++ packages/core/src/react/runtimes/cloud/useCloudThreadListAdapter.tsx
@@ -1,257 +1,62 @@
-declare const process: { env: Record<string, string | undefined> };
-
 import {
   type FC,
   type PropsWithChildren,
-  type RefObject,
   useCallback,
-  useEffect,
   useMemo,
   useRef,
-  useState,
 } from "react";
-import { AssistantCloud } from "assistant-cloud";
 import type { RemoteThreadListAdapter } from "../../../runtimes/remote-thread-list/types";
-import { InMemoryThreadListAdapter } from "../../../runtimes/remote-thread-list/adapter/in-memory";
-import { useAssistantCloudThreadHistoryAdapter } from "./AssistantCloudThreadHistoryAdapter";
 import {
   RuntimeAdapterProvider,
   type RuntimeAdapters,
 } from "../RuntimeAdapterProvider";
-import { CloudFileAttachmentAdapter } from "./CloudFileAttachmentAdapter";
-import { isRecord } from "../../../utils/json/is-json";
-
-type ThreadData = {
-  externalId: string | undefined;
-};
-
-type CloudThreadListAdapterOptions = {
-  cloud?: AssistantCloud | undefined;
-
-  create?: (() => Promise<ThreadData>) | undefined;
-  delete?: ((threadId: string) => Promise<void>) | undefined;
-};
-
-const toCustom = (value: unknown): Record<string, unknown> | undefined =>
-  isRecord(value) ? value : undefined;
-
-const baseUrl =
-  typeof process !== "undefined" &&
-  process?.env?.NEXT_PUBLIC_ASSISTANT_BASE_URL;
-const autoCloud = baseUrl
-  ? new AssistantCloud({ baseUrl, anonymous: true })
-  : undefined;
-
-const CLOUD_THREAD_PAGE_SIZE = 20;
-
-type CloudListCursor = {
-  activeCursor: string | undefined;
-  archivedCursor: string | undefined;
-  activeExhausted: boolean;
-  archivedExhausted: boolean;
-};
-
-const parseListCursor = (after: string | undefined): CloudListCursor => {
-  const fallback: CloudListCursor = {
-    activeCursor: after,
-    archivedCursor: undefined,
-    activeExhausted: false,
-    archivedExhausted: false,
-  };
-  if (!after || !after.startsWith("{")) return fallback;
-  try {
-    const parsed = JSON.parse(after);
-    if (!isRecord(parsed)) return fallback;
-    return {
-      activeCursor: typeof parsed.a === "string" ? parsed.a : undefined,
-      archivedCursor: typeof parsed.r === "string" ? parsed.r : undefined,
-      activeExhausted: parsed.ae === true,
-      archivedExhausted: parsed.re === true,
-    };
-  } catch {
-    return fallback;
-  }
-};
-
-const useCloudThreadListAdapters = (
-  adapterRef: RefObject<CloudThreadListAdapterOptions>,
-): RuntimeAdapters => {
-  const history = useAssistantCloudThreadHistoryAdapter({
-    get current() {
-      return adapterRef.current.cloud ?? autoCloud!;
-    },
-  });
-  const [attachments] = useState(
-    () =>
-      new CloudFileAttachmentAdapter(
-        () => adapterRef.current.cloud ?? autoCloud!,
-      ),
-  );
-  return useMemo(
-    () => ({
-      history,
-      attachments,
-    }),
-    [history, attachments],
-  );
-};
+import {
+  autoCloud,
+  createCloudThreadListAdapter,
+  type CloudThreadListAdapterOptions,
+} from "./createCloudThreadListAdapter";
 
 export const useCloudThreadListAdapter = (
   adapter: CloudThreadListAdapterOptions,
 ): RemoteThreadListAdapter => {
+  // Synced during render, not in an effect: the memo below recreates the
+  // adapter in the same render a cloud swap arrives, and the factory pins the
+  // cloud instance it reads at creation.
   const adapterRef = useRef(adapter);
-  useEffect(() => {
-    adapterRef.current = adapter;
-  }, [adapter]);
+  adapterRef.current = adapter;
+
+  const cloud = adapter.cloud ?? autoCloud;
+  const base = useMemo(
+    () => createCloudThreadListAdapter(() => adapterRef.current),
+    // oxlint-disable-next-line react-hooks/exhaustive-deps -- the factory branches on the cloud instance; other options are read through the ref
+    [cloud],
+  );
+
+  const baseRef = useRef(base);
+  baseRef.current = base;
 
   const unstable_useAdapters = useCallback(function useCloudAdapters() {
-    return useCloudThreadListAdapters(adapterRef);
+    return baseRef.current.unstable_useAdapters!() as RuntimeAdapters;
   }, []);
 
   const unstable_Provider = useCallback<FC<PropsWithChildren>>(
     function Provider({ children }) {
-      const adapters = useCloudThreadListAdapters(adapterRef);
+      const adapters = unstable_useAdapters();
       return (
         <RuntimeAdapterProvider adapters={adapters}>
           {children}
         </RuntimeAdapterProvider>
       );
     },
-    [],
+    [unstable_useAdapters],
   );
 
-  const cloud = adapter.cloud ?? autoCloud;
   return useMemo<RemoteThreadListAdapter>(() => {
-    if (!cloud) {
-      const ref = adapterRef;
-      const inMemory = new InMemoryThreadListAdapter();
-      inMemory.initialize = async (threadId: string) => {
-        const result = await ref.current.create?.();
-        return { remoteId: threadId, externalId: result?.externalId };
-      };
-      return inMemory;
-    }
-
+    if (base.unstable_useAdapters === undefined) return base;
     return {
-      list: async ({ after } = {}) => {
-        const {
-          activeCursor,
-          archivedCursor,
-          activeExhausted,
-          archivedExhausted,
-        } = parseListCursor(after);
-        const [{ threads: activeThreads }, { threads: archivedThreads }] =
-          await Promise.all([
-            activeExhausted
-              ? Promise.resolve({ threads: [] })
-              : cloud.threads.list({
-                  limit: CLOUD_THREAD_PAGE_SIZE,
-                  ...(activeCursor ? { after: activeCursor } : {}),
-                }),
-            archivedExhausted
-              ? Promise.resolve({ threads: [] })
-              : cloud.threads.list({
-                  is_archived: true,
-                  limit: CLOUD_THREAD_PAGE_SIZE,
-                  ...(archivedCursor ? { after: archivedCursor } : {}),
-                }),
-          ]);
-        const activeNext =
-          !activeExhausted && activeThreads.length === CLOUD_THREAD_PAGE_SIZE
-            ? activeThreads.at(-1)?.id
-            : undefined;
-        const archivedNext =
-          !archivedExhausted &&
-          archivedThreads.length === CLOUD_THREAD_PAGE_SIZE
-            ? archivedThreads.at(-1)?.id
-            : undefined;
-        const threads = [...activeThreads, ...archivedThreads];
-        return {
-          threads: threads.map((t) => ({
-            status: t.is_archived ? "archived" : "regular",
-            remoteId: t.id,
-            title: t.title,
-            lastMessageAt: t.last_message_at
-              ? new Date(t.last_message_at)
-              : undefined,
-            externalId: t.external_id ?? undefined,
-            custom: toCustom(t.metadata),
-          })),
-          nextCursor:
-            activeNext || archivedNext
-              ? JSON.stringify({
-                  a: activeNext,
-                  r: archivedNext,
-                  ...(activeNext === undefined ? { ae: true } : {}),
-                  ...(archivedNext === undefined ? { re: true } : {}),
-                })
-              : undefined,
-        };
-      },
-
-      initialize: async () => {
-        const createTask = adapterRef.current.create?.() ?? Promise.resolve();
-        const t = await createTask;
-        const external_id = t ? t.externalId : undefined;
-        const { thread_id: remoteId } = await cloud.threads.create({
-          last_message_at: new Date(),
-          external_id,
-        });
-
-        return { externalId: external_id, remoteId: remoteId };
-      },
-
-      rename: async (threadId, newTitle) => {
-        return cloud.threads.update(threadId, { title: newTitle });
-      },
-      updateCustom: async (threadId, custom) => {
-        return cloud.threads.update(threadId, { metadata: custom ?? null });
-      },
-      archive: async (threadId) => {
-        return cloud.threads.update(threadId, { is_archived: true });
-      },
-      unarchive: async (threadId) => {
-        return cloud.threads.update(threadId, { is_archived: false });
-      },
-      delete: async (threadId) => {
-        await adapterRef.current.delete?.(threadId);
-        return cloud.threads.delete(threadId);
-      },
-
-      generateTitle: async (threadId, messages) => {
-        // Filter messages to only include content types the title generator understands
-        // (reasoning, source, etc. are not needed for title generation)
-        // TODO serialize these to a more efficient format
-        const filteredMessages = messages.map((msg) => ({
-          ...msg,
-          content: msg.content.filter(
-            (part) => part.type === "text" || part.type === "tool-call",
-          ),
-        }));
-
-        return cloud.runs.stream({
-          thread_id: threadId,
-          assistant_id: "system/thread_title",
-          messages: filteredMessages,
-        });
-      },
-
-      fetch: async (threadId: string) => {
-        const thread = await cloud.threads.get(threadId);
-        return {
-          status: thread.is_archived ? "archived" : "regular",
-          remoteId: thread.id,
-          title: thread.title,
-          lastMessageAt: thread.last_message_at
-            ? new Date(thread.last_message_at)
-            : undefined,
-          externalId: thread.external_id ?? undefined,
-          custom: toCustom(thread.metadata),
-        };
-      },
-
+      ...base,
       unstable_Provider,
       unstable_useAdapters,
     };
-  }, [cloud, unstable_Provider, unstable_useAdapters]);
+  }, [base, unstable_Provider, unstable_useAdapters]);
 };
