+++ packages/core/src/runtimes/remote-thread-list/title.ts
@@ -1,0 +1,3 @@
+export const isTitleSourceMessage = (message: {
+  status?: { type: string } | undefined;
+}) => message.status?.type !== "running";
