+++ packages/store/src/client.ts
@@ -12,8 +12,10 @@ export {
 export {
   DefaultAssistantClient,
   useAssistantContextProvider,
+  useAssistantContextValue,
 } from "./utils/react-assistant-context";
 export { useConfiguredAui } from "./useAui";
+export { getClientId } from "./utils/client-accessor";
 export { getProxiedAssistantState } from "./utils/proxied-assistant-state";
 export {
   useAssistantClientRef,
